# backend/openapi/unital-v1.yaml
openapi: 3.1.0
info:
  title: Unital API
  version: 1.0.0
  description: |
    Building & complex management platform. Consolidated gateway API for the
    Unital Go microservices (identity, property, billing, facilities,
    operations, notifications/communications).

    **Auth**: `POST /auth/login` returns an access token (15 min) + refresh
    token. Send `Authorization: Bearer <access>` on every other call.

    **Errors**: all failures use RFC 9457 problem details with a
    machine-readable `code` and `trace_id`.
  contact: { name: Unital, url: https://unital.app }
servers:
  - url: http://localhost:8080/api/v1
    description: local gateway
  - url: http://localhost:18080/api/v1
    description: local gateway (dev port)
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
            example:
              email: sara@example.com
              password: secret12
              full_name: سارا احمدی
              phone: "09120000001"
              role: owner
      responses:
        "201":
          description: Account created (unverified — login requires /auth/verify)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email + password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Token pair
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401": { $ref: "#/components/responses/Problem" }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke a refresh token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: { type: string }
      responses:
        "200": { description: Logged out }
  /auth/verify:
    post:
      tags: [Auth]
      summary: Verify account with emailed token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "200": { description: Verified }
        "401": { $ref: "#/components/responses/Problem" }
  /me:
    get:
      tags: [Auth]
      summary: My profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    patch:
      tags: [Auth]
      summary: Update my profile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name: { type: string }
                phone: { type: string }
                national_code: { type: string }
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
...
  /buildings/{buildingID}/memberships:
    parameters:
      - $ref: "#/components/parameters/BuildingID"
    post:
      tags: [Memberships]
      summary: Grant a building role (manager only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, role]
              properties:
                user_id: { type: string }
                role:
                  type: string
                  enum: [manager, board_member, staff, owner, resident]
      responses:
        "201":
          description: Granted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Membership" }
        "403": { $ref: "#/components/responses/Problem" }
    get:
      tags: [Memberships]
      summary: List building memberships (members only)
      responses:
        "200":
          description: Membership list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Membership" }
  /buildings/{buildingID}/memberships/{membershipID}:
    delete:
      tags: [Memberships]
      summary: Revoke a membership (manager only)
      parameters:
        - $ref: "#/components/parameters/BuildingID"
        - name: membershipID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Revoked }
        "403": { $ref: "#/components/responses/Problem" }

  # ---------------- Property ----------------
...
  /auth/password-reset:
    post:
      tags: [Auth]
      summary: Request a password-reset email (always 200 — never reveals accounts)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200": { description: Reset requested if the account exists }
  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      summary: Reset the password with the emailed token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token: { type: string }
                new_password: { type: string }
      responses:
        "200": { description: Password updated }
        "401": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BuildingID:
      name: buildingID
      in: path
      required: true
      schema: { type: string }
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details with machine-readable code
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string, example: UNIT_NOT_IN_BUILDING }
        detail: { type: string }
        trace_id: { type: string }
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password:
          type: string
          description: "≥8 chars with a letter and a digit"
        full_name: { type: string }
        phone: { type: string }
        role:
          type: string
          enum: [manager, staff, owner, resident, board_member]
          default: resident
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: Bearer }
        expires_in: { type: integer, example: 900 }
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        full_name: { type: string }
        phone: { type: string }
        national_code: { type: string }
        platform_role: { type: string }
        email_verified: { type: boolean }
        created_at: { type: string, format: date-time }
    Membership:
      type: object
      properties:
        id: { type: string }
        user_id: { type: string }
        building_id: { type: string }
        role: { type: string, enum: [manager, board_member, staff, owner, resident] }
        from: { type: string, format: date-time }
...